正解画面、ダイアログの調整等
