+++ src/app/home/question1/nottori-ans.tsx
@@ -40,7 +40,7 @@ const Nottori = () => {
   const steps = [
     {
       title: "解答について",
-      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ右下の「間違い無し」ボタンを押してください。問題がないメールの場合もあります。",
+      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ画面下の「まちがいなし」ボタンを押してください。問題がないメールの場合もあります。",
       target: buttonRef,
     }
   ]
@@ -194,7 +194,7 @@ const Nottori = () => {
           <Footer />
         </View>
       </ScrollView>
-      <FooterButton label="まちがいなし" onPress={handlePress1} />
+      <FooterButton ref={buttonRef} label="まちがいなし" onPress={handlePress1} />
       <Modal
         visible={modalVisible}
         transparent={true}
