+++ src/app/home/question3/quiz1-ans.tsx
@@ -31,7 +31,7 @@ const quiz1 = () => {
   const steps = [
     {
       title: "解答について",
-      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ右下の「間違い無し」ボタンを押してください。問題がないメールの場合もあります。",
+      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ画面下の「まちがいなし」ボタンを押してください。問題がないメールの場合もあります。",
       target: buttonRef,
     }
   ]
@@ -153,7 +153,7 @@ const quiz1 = () => {
         </View>
 
       </ScrollView>
-      <FooterButton label="まちがいなし" onPress={handlepressCorrect} />
+      <FooterButton ref={buttonRef} label="まちがいなし" onPress={handlepressCorrect} />
       <Modal
         visible={modalVisible}
         transparent={true}
