+++ src/app/home/question5/quiz2-another.tsx
@@ -25,7 +25,7 @@ const Quiz2 = () => {
   const steps = [
     {
       title: "問題について",
-      description: "問題の回答をスタートするには！ボタンを押します。",
+      description: "問題の回答をスタートするには「回答する」ボタンを押します。",
       target: buttonRef,
     }
   ]
