+++ src/app/home/question7/rightspot-ans.tsx
@@ -28,7 +28,7 @@ export default function App() {
   const steps = [
     {
       title: "解答について",
-      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ右下の「間違い無し」ボタンを押してください。問題がないメールの場合もあります。",
+      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ画面下の「まちがいなし」ボタンを押してください。問題がないメールの場合もあります。",
       target: buttonRef,
     }
   ]
@@ -141,7 +141,7 @@ export default function App() {
       </ScrollView>
       <TouchableOpacity onPress={handlepressIncorrect}>
       </TouchableOpacity>
-      <FooterButton label="まちがいなし" onPress={handlepressCorrect} />
+      <FooterButton ref={buttonRef} label="まちがいなし" onPress={handlepressCorrect} />
       <Modal
         visible={modalVisible}
         transparent={true}
