+++ src/app/home/question7/rightspot.tsx
@@ -23,7 +23,7 @@ export default function App() {
   const steps = [
     {
       title: "問題について",
-      description: "問題の回答をスタートするには！ボタンを押します。",
+      description: "問題の回答をスタートするには「回答する」ボタンを押します。",
       target: buttonRef,
     }
   ]
