+++ src/app/home/question8/goji-ans.tsx
@@ -30,7 +30,7 @@ export default function App() {
   const steps = [
     {
       title: "解答について",
-      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ右下の「間違い無し」ボタンを押してください。問題がないメールの場合もあります。",
+      description: "画面に表示されているメールには詐欺メールに書かれているような表現が含まれている可能性があります。\n怪しいと感じる表現が含まれる部分をタップして解答してください。\n特に怪しいと感じる部分がなければ画面下の「まちがいなし」ボタンを押してください。問題がないメールの場合もあります。",
       target: buttonRef,
     }
   ]
@@ -167,7 +167,7 @@ export default function App() {
           </TouchableOpacity>
         </View>
       </ScrollView>
-      <FooterButton label="まちがいなし" onPress={handlePressHome} />
+      <FooterButton ref={buttonRef} label="まちがいなし" onPress={handlePressHome} />
       <Modal
         visible={modalVisible}
         transparent={true}
